requirements.txt
requests
websockets
python-dateutil

google-cloud-secret-manager
google-cloud-pubsub
